+++ programs/sll_partition.py
@@ -1,37 +1,11 @@
 # in a sll make a partition to around a value x. All nodes less than x  come before that all nodes equal or greater to x
 # the x value can appear any where in the right partition. x not needs to apear in the middle, just in right partition.
+from single_linked_lists import Node
 
-
-class Node:
-    def __init__(self, data):
-        self.data = data
-        self.next = None
-        self.head = None
-
-    def display(self):
-        if self.head is None:
-            print("Linked List is empty")
-        else:
-            temp = self.head
-            while temp:
-                print(temp.data, end="  ")
-                if temp.next:
-                    print("--->", end=" ")
-                temp = temp.next
-            print()
-
-
-for i in range(int(input("How many nodes? "))):
-    nn = Node(int(input(f"Node {i} : ")))
-    if i == 0:
-        n = nn
-        h = n
-        h.head = n
-    n.next = nn
-    n = n.next
-
-
+h = Node(None)
+h = h.list_builder()
 h.display()
+
 x = int(input("Which element wants to be the partition reference: "))
 
 p1 = h.head  # p1 = pointer 1
@@ -42,9 +16,8 @@ while p1.next:  # p1 walk sll
         continue
     p2 = p1.next  # p2 = poniter 1
     while p2:  # p2 walk sll
-        if (
-            p2.data < p1.data and p2.data < x
-        ):  # compare p1 and p2 and if necesary switch them
+        # compare p1 and p2 and if necesary switch them
+        if p2.data < p1.data and p2.data < x:
             temp = p1.data
             p1.data = p2.data
             p2.data = temp
